alot

+++ crates/feagi-data/src/index_range_managers/feagi_index_range_manager_error.rs
@@ -13,7 +13,7 @@ pub struct FeagiIndexManagerLimit {
 #[derive(FeagiErrorKey)]
 pub struct FeagiIndexManagerInvalidIndex {
     context: &'static str,
-    index: u32
+    index: usize
 }
 
 #[derive(FeagiErrorKey)]
